Rename library's use_cases/ to commands/
Every read in the module now lives in `queries/`, so the package takes
the ADR's end-state name, symmetric with `queries/`. Classes keep the
`*UseCase` suffix: "use case" is the layer role and reads are use cases
too, while the package encodes the command/query kind.

The `queries-are-dead-ends` contract already covers both
`src.application.*.use_cases` and `src.application.*.commands`, so no
pyproject change is needed — `reading` and `identity` are the only
modules still on the old name.
