+++ backend/src/application/library/commands/book_files/ebook_deletion_use_case.py
