+++ backend/src/application/library/commands/book_management/__init__.py
