+++ backend/src/application/library/commands/book_management/create_book_use_case.py
