+++ backend/src/containers/library.py
@@ -1,5 +1,21 @@
 from dependency_injector import containers, providers
 
+from src.application.library.commands.book_files.ebook_deletion_use_case import (
+    EbookDeletionUseCase,
+)
+from src.application.library.commands.book_files.ebook_upload_use_case import EbookUploadUseCase
+from src.application.library.commands.book_management.create_book_use_case import (
+    CreateBookUseCase,
+)
+from src.application.library.commands.book_management.delete_book_use_case import (
+    DeleteBookUseCase,
+)
+from src.application.library.commands.book_management.mark_book_viewed_use_case import (
+    MarkBookViewedUseCase,
+)
+from src.application.library.commands.book_management.update_reading_stage_use_case import (
+    UpdateReadingStageUseCase,
+)
 from src.application.library.queries.get_book_details_use_case import GetBookDetailsUseCase
 from src.application.library.queries.get_books_with_counts_use_case import (
     GetBooksWithCountsUseCase,
@@ -9,22 +25,6 @@ from src.application.library.queries.get_ereader_metadata_use_case import (
 )
 from src.application.library.queries.get_recently_viewed_books_use_case import (
     GetRecentlyViewedBooksUseCase,
-)
-from src.application.library.use_cases.book_files.ebook_deletion_use_case import (
-    EbookDeletionUseCase,
-)
-from src.application.library.use_cases.book_files.ebook_upload_use_case import EbookUploadUseCase
-from src.application.library.use_cases.book_management.create_book_use_case import (
-    CreateBookUseCase,
-)
-from src.application.library.use_cases.book_management.delete_book_use_case import (
-    DeleteBookUseCase,
-)
-from src.application.library.use_cases.book_management.mark_book_viewed_use_case import (
-    MarkBookViewedUseCase,
-)
-from src.application.library.use_cases.book_management.update_reading_stage_use_case import (
-    UpdateReadingStageUseCase,
 )
